+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Karafka Test gem changelog
 
 ## 2.3.1 (2024-03-07)
+- [Enhancement] Prevent usage of testing when Karafka is not loaded.
+- [Enhancement] Prevent usage of testing when Karafka is not initialized.
 - [Fix] Same CG multiplexing topic builder fails with `TopicInManyConsumerGroupsError`.
 
 ## 2.3.0 (2024-01-26)